Add hack to switch file type to PNG
PDX materials are stored as .dds, however for my purposes I've run
some processes that move the files to .png. This hack allows me to
load the files directly without having to manually change a bunch of
things, assuming the files are in the same folder path as the original.